### 2025-07-07 f27f534
add panel labels to `Gallery`
