### 2025-07-09 #8544 @zkbt
Hi @warrickball , thanks so much! Sorry about the long delay. Whether or not it'd help speed things along, I've been meaning to offer to help review for JOSS, so please feel free to add me to your list of candidate reviewers!

Yes, I was hoping to do this as a AAS Journal collaboration. I also submitted a paper that makes heavy use of `exoatlas` to the AAS Journals at the same time: https://arxiv.org/abs/2507.02136
