### 2025-07-09 #8544 @warrickball
> Whether or not it'd help speed things along, I've been meaning to offer to help review for JOSS, so please feel free to add me to your list of candidate reviewers!

That's great! You can add yourself to our reviewer database by [signing up on its website](https://reviewers.joss.theoj.org/).

> Yes, I was hoping to do this as a AAS Journal collaboration. I also submitted a paper that makes heavy use of exoatlas to the AAS Journals at the same time: https://arxiv.org/abs/2507.02136

I've not done this myself so am not 100% sure of the process but this should be an explicit choice in the AAS Journals submission. @augustfly, can you confirm the AAS Journals submission has been associated with the JOSS software review?
