### 2025-07-09 #8544 @augustfly
Thanks, I can associate it now @warrickball and I'll add it to the AAS/JOSS tracking sheet.

@zkbt I'll write up a data review so you know what we expect in terms of cross-citing of the ApJL and JOSS articles. You'll get that information when you receive the first peer-review report.

thanks
