### 2025-08-05 cde5c1e
add Mamajek interpolator for converting main-sequence stellar quantities
