+++ exoatlas/models/__init__.py
@@ -2,3 +2,4 @@ from .seager import *
 from .zeng import *
 from .kopparapu import *
 from .chen import *
+from .mamajek import *
