+++ exoatlas/models/relation.py
@@ -1,0 +1,216 @@
+# imports from general utilities
+import os
+import numpy as np
+import matplotlib.pyplot as plt
+import scipy.interpolate
+import astropy.io.ascii
+import pkg_resources
+import string
+import warnings
+
+
+def remove_punctuation(s):
+    """
+    This helper function cleans a string of all punctuation.
+    """
+    translator = s.maketrans("", "", " " + string.punctuation)
+    return s.translate(translator)
+
+
+class Relation:
+    """
+    Base class for astrophysical relations,
+    defining tools to read tables, define methods, etc...
+    """
+
+    def __init__(self, filename, **kwargs):
+        """
+        Initialize a Relation object.
+        """
+
+        # make sure the basic settings are included
+        self.name = "Some General Relation"
+        self.bibcode = "(no ADS bibcode defined)"
+
+        # store the data filename
+        self.filename = filename
+
+        # load the data table
+        self.load(**kwargs)
+
+    def load(self, **kwargs):
+        """
+        Load the data table, and store it internally to this object.
+        """
+
+        # figure out the path to the data file, relative to this package
+        # path = os.path.dirname(__file__) + '/'+ self.filename
+        path = pkg_resources.resource_filename(__name__, self.filename)
+        # print("loading data from {0}".format(path))
+
+        # load as an astropy table
+        self.table = astropy.io.ascii.read(
+            path, fill_values=[("...", np.nan)], **kwargs
+        )
+
+        # give an update
+        # print("   ...success!")
+
+    @property
+    def possible(self):
+        """
+        What are the possible columns in this relation?
+        """
+        try:
+            # if custom columns are defined, use those
+            return self.columns
+        except AttributeError:
+            # otherwise, use all the columns in the table
+            return self.table.keys()
+
+    def limits(self, key):
+        """
+        Return the range of values spanned by
+        a particular aspect of the relation.
+        """
+        return np.nanmin(self.table[key]), np.nanmax(self.table[key])
+
+    def describe(self, key):
+        """Describe one column in the table."""
+        # print("{} = {}".format(key, self.descriptions[key]))
+
+    def summarize(self):
+        """Summarize all the possible columns in the table."""
+        # print("")
+        # print("The columns in {} are:".format(self.name))
+        # for key in self.possible:
+        # print("{:>20} = {}".format(key, self.descriptions.get(key, "???")))
+
+    def tofrom(self, outkey, verbose=True):
+        """
+        Return a function that returns a function
+        that will interpolate the values for given output (`outkey`),
+        for some specified input (`inkey`).
+
+        For example
+
+            interpolator = Mamajek.tofrom('BCv')('Teff')
+
+        will return a function called interpolator
+        that takes values for Teff as input,
+        and returns values for BCv.
+        """
+
+        # create a function that takes one input as a keyword arg
+        def function(inkey):
+            return self.interpolator(inkey=inkey, outkey=outkey)
+
+        return function
+
+    def interpolator(self, inkey=None, outkey=None):
+        """
+        Create an interpolator function, going from inkey to outkey.
+        """
+        # print("creating interpolator to convert {0} to {1}".format(inkey, outkey))
+        try:
+            x = self.table[inkey]
+        except:
+            warnings.warn(
+                "it seems like the attempted input key {0} isn't valid".format(inkey)
+            )
+            return None
+        try:
+            y = self.table[outkey]
+        except:
+            warnings.warn(
[…]
