+++ exoatlas/models/data/mamajek_dwarfproperties.txt
@@ -1,0 +1,141 @@
+# "A Modern Mean Dwarf Stellar Color and Effective Temperature Sequence"
+# http://www.pas.rochester.edu/~emamajek/EEM_dwarf_UBVIJHK_colors_Teff.txt
+# Eric Mamajek
+# Version 2022.04.16
+#
+# Much of the original content of this table was incorporated into
+# Table 5 of Pecaut & Mamajek (2013, ApJS, 208, 9;
+# http://adsabs.harvard.edu/abs/2013ApJS..208nan9P), so that
+# reference should be cited until an updated version of the table is
+# published.  Parts of this table for A/F/G stars also appeared in
+# Table 3 of Pecaut, Mamajek, & Bubar (2012, ApJ 756, 154).  For
+# example, Table 5 of Pecaut & Mamajek did not include the absolute
+# magnitude and luminosity estimates, nor the properties of dwarfs of
+# types L, T, Y, or O3-O8.
+# The author's notes on standard stars and mean parameters for stars
+# for each dwarf spectral type can be found in the directory at:
+# http://www.pas.rochester.edu/~emamajek/spt/. See further notes and
+# caveats after the table. Colors are based on Johnson UBV, Tycho
+# BtVt, Cousins RcIc, Gaia DR2 G/Bp/Rp, Sloan izY, 2MASS JHKs,
+# MKO JK (for Y dwarfs), WISE W1/W2/W3/W4 photometry.
+# G-V color is Gaia DR2 G - Johnson V.
+#
+SpT   Teff  logT   BCv    logL   Mbol R_Rsun   Mv    B-V    Bt-Vt  G-V    Bp-Rp  G-Rp   M_G    b-y    U-B    V-Rc   V-Ic   V-Ks   J-H    H-Ks   M_J    M_Ks   Ks-W1   W1-W2  W1-W3  W1-W4   g-r   i-z  z-Y  Msun  SpT2  
+O3V    44900 4.652 -4.01   5.82  -9.81 13.43   -5.80 -0.330  nan    nan    nan    nan    nan    nan -1.175  nan    nan  nan  nan  nan  nan   nan   nan     nan    nan    nan     nan   nan  nan  59    O3V   
+O4V    42900 4.632 -3.89   5.65  -9.39 12.13   -5.50 -0.326  nan    nan    nan    nan    nan    nan -1.160  nan    nan  nan  nan  nan  nan   nan   nan     nan    nan    nan     nan   nan  nan  48    O4V   
+O5V    41400 4.617 -3.76   5.54  -9.11 11.45   -5.35 -0.323  nan    nan    nan    nan    nan   -0.133 -1.150  nan    nan  nan  nan  nan  nan   nan   nan     nan    nan    nan    -0.620 nan  nan  43    O5V   
+O5.5V  40500 4.607 -3.67   5.44  -8.87 10.71   -5.20 -0.322  nan    nan    nan    nan    nan   -0.133 -1.145  nan    nan  nan  nan  nan  nan   nan   nan     nan    nan    nan    -0.620 nan  nan  38    O5.5V 
[…]
