### 2025-09-08 e1cae80
add `sliceaxis` to `Map` to allow filtering populations
