+++ exoatlas/tests/test_calculations.py
@@ -15,8 +15,8 @@ eccentricity,
 argument_of_periastron,
 transit_impact_parameter,
 insolation,
-relative_insolation,
-log_relative_insolation,
+relative_instellation,
+log_relative_instellation,
 relative_cumulative_xuv_insolation,
 teq,
 planet_luminosity,
