### 2025-09-17 4fb9096
fix bug where annotation wasn't propagating into subpopulations or slices
