### 2025-09-18 ac1a072
incorporate slicing option into all `Maps`; add units to many `Plottable` limits
