+++ exoatlas/visualizations/plottables/preset_plottables.py
@@ -14,7 +14,7 @@ class Flux(Plottable):
 class LogRelativeInstellation(Plottable):
     source = "relative_instellation"
     label = "log10[Planet Bolometric Flux Received\n(relative to Earth)]"
-    scale = "log"
+    scale = "linear"
     lim = [4, -4]
     symbol = r"$\sf \log_{10}(f/f_\oplus)$"
 
@@ -22,7 +22,7 @@ class LogRelativeInstellation(Plottable):
 class Teq(Plottable):
     source = "teq"
     scale = "log"
-    lim = [200, 2000]
+    lim = [200, 2000] * u.K
 
     def _update_label(self):
 
@@ -48,35 +48,36 @@ class Radius(Plottable):
     source = "radius"
     label = "Planet Radius (Earth radii)"
     scale = "log"
-    lim = [0.3, 30]
+    lim = [0.3, 30] * u.R_earth
+    symbol = "$\sf R_p$"
 
 
 class LogRadius(Plottable):
     source = "log_relative_escape_velocity"
     scale = "log"
     lim = [-1, 1]
-    symbol = r"$\sf \log_{10}(v_{\rm esc}/v_{esc,\oplus})$"
+    symbol = r"$\sf \log_{10}(R_p/R_\oplus)$"
 
 
 class Mass(Plottable):
     source = "mass"
     label = "Planet Mass\n(Earth masses)"
     scale = "log"
-    lim = [0.03, 3000]
+    lim = [0.03, 3000] * u.M_earth
 
 
 class SemimajorAxis(Plottable):
     source = "semimajoraxis"
     label = "Semimajor Axis (AU)\n"
     scale = "log"
-    lim = [0.001, 1000]
+    lim = [0.001, 1000] * u.AU
 
 
 class AngularSeparation(Plottable):
     source = "angular_separation"
     label = "Angular Separation (arcsec)\n"
     scale = "log"
-    lim = [0.001, 10]
+    lim = [0.001, 10] * u.arcsec
 
 
 class Contrast(Plottable):
@@ -90,14 +91,14 @@ class KludgedMass(Plottable):
     source = "kludge_mass"
     label = "Planet Mass or msini\n(Earth masses)"
     scale = "log"
-    lim = [0.03, 4131]
+    lim = [0.03, 4131] * u.M_earth
 
 
 class StellarTeff(Plottable):
     source = "stellar_teff"
     label = "Stellar Temperature (K)"
     scale = "linear"
-    lim = [2000, 7000]
+    lim = [2000, 7000] * u.K
 
 
 class StellarLuminosity(Plottable):
@@ -110,7 +111,7 @@ class StellarLuminosity(Plottable):
 
 class LogRelativeStellarLuminosity(Plottable):
     source = "log_relative_stellar_luminosity"
-    scale = "log"
+    scale = "linear"
     lim = [-4, 1]
     symbol = r"$\sf \log_{10}(L_\star/L_\odot)$"
 
@@ -119,21 +120,16 @@ class Distance(Plottable):
     source = "distance"
     label = "Distance\n(parsecs)"
     scale = "log"
-    lim = [5, 1000]
-
-
-class Distance(Plottable):
-    source = "distance"
-    label = "Distance\n(parsecs)"
-    scale = "log"
-    lim = [5, 1000]
+    lim = [5, 1000] * u.pc
+    symbol = r"$\sf D$"
 
 
 class EscapeVelocity(Plottable):
     source = "escape_velocity"
     label = "Escape Velocity\n(km/s)"
     scale = "log"
-    lim = [2, 500]
+    lim = [2, 500] * u.km / u.s
[…]
