### 2025-09-28 86912c7
implement most of shoreline plotting with posteriors
