+++ exoatlas/visualizations/shorelines/__init__.py
@@ -1,0 +1,2 @@
+from .shoreline import *
+from .preset_shorelines import *
