+++ exoatlas/visualizations/__init__.py
@@ -1,2 +1,3 @@
 from .maps import *
 from .galleries import *
+from .shorelines import *
