+++ exoatlas/visualizations/shorelines/preset_shorelines.py
@@ -1,0 +1,365 @@
+from .shoreline import *
+from ..animation import get_animation_writer
+from ..plottables import *
+from ...models import *
+
+
+class Shoreline_SemimajorAxis_x_StellarLuminosity_x_Radius(ShorelineErrorMap):
+    """ """
+
+    xaxis = SemimajorAxis(lim=[1e-3, 1e1])
+    yaxis = StellarLuminosity(lim=[1e-4, 1e1])
+    sliceaxis = Radius(lim=[0.3, 2.0] * u.Rearth)
+
+    def _convert_radius_to_relative_escape_velocity(self, radius):
+        """
+        Use a mass-radius relation to convert from
+        log(radius) to log(relative escape velocity).
+        """
+        lnR = np.log(radius.to_value(u.Rearth))
+        mass_radius_slope = 3.353
+        mass_radius_intercept = 0.011
+        lnM = mass_radius_slope * lnR + mass_radius_intercept
+        lnV = 0.5 * (lnM - lnR)
+        return np.exp(lnV)
+
+    def _get_slice_relative_escape_velocities(self, N=100):
+        """
+        Generate log10(relative escape velocity) for this radius slice,
+        including both one value for the "center" of the slice
+
+        Parameters
+        ----------
+        N : int
+            The number of samples to generate.
+
+        Returns
+        -------
+        log10v : float
+            A grid of log10(relative escape velocity) values
+            that span the slice from limit to limit.
+        """
+
+        # uniform grid in R
+        # R = np.linspace(*self.plottable['slice'].lim, num=N)
+
+        # uniform grid in log R
+        lnR = np.linspace(
+            *np.log(self.plottable["slice"].lim.to_value(u.R_earth)), num=N
+        )
+        R = np.exp(lnR) * u.R_earth
+
+        # use mass-radius relation to convert to escape velocity
+        relative_escape_velocity = self._convert_radius_to_relative_escape_velocity(R)
+        return relative_escape_velocity
+
+    def plot_hz_earth(self, color="seagreen", alpha=0.75, ax=None, **kw):
+        """
+        Plot the Earth equivalent habitable zone.
+
+        Parameters
+        ----------
+        **kw : dict
+            All keywords will be passed to `plt.plot`
+        """
+
+        plt.sca(ax or self.ax)
+
+        # create grid of luminosities
+        L = np.logspace(-4, 2) * u.Lsun
+
+        # plot Earth-equivalent instellation distance
+        a_ee = 1 * u.AU * (L / u.Lsun) ** 0.5
+        plt.plot(a_ee, L, color=color, alpha=alpha, **kw)
+
+    def plot_hz_kopparapu(
+        self, color="seagreen", alpha=0.25, linewidth=0, zorder=-100, ax=None, **kw
+    ):
+        """
+        Plot the Kopparapu et al. (2013) habitable zone.
+
+        This includes the Kopparapu et al. (2013)
+        HZ as a function of stellar Teff (translated
+        from luminosity via the Mamajek table).
+
+        It makes no use of the planet's mass,
+        just the stellar effective temperature.
+
+        Parameters
+        ----------
+        **kw : dict
+            All keywords will be passed to `plt.fill_betweenx`
+        """
+
+        plt.sca(ax or self.ax)
+
+        # create grid of luminosities
+        L = np.logspace(-4, 2) * u.Lsun
+
+        # define the HZ functions from Kopparapu
+        f_inner = make_hz("moist-greenhouse")
+        f_outer = make_hz("maximum-greenhouse")
+
+        # translate from luminosity to Teff
+        m = Mamajek()
[…]

[…]
