### 2025-10-12 7cb32df
add shoreline for plots and `probability_of_atmosphere` calculation to populations
