+++ exoatlas/visualizations/shorelines/__init__.py
@@ -1,2 +1,1 @@
-from .shoreline import *
 from .preset_shorelines import *
