### 2025-11-26 db37c04
minor visualization updates; add very rough twinkling kludge
