+++ exoatlas/telescopes/buckets.py
@@ -83,6 +83,31 @@ def define_JWST_unit(wavelength=5 * u.micron, efficiency=0.3, **kw):
         **kw,
     )
 
+def define_GMOS_unit(wavelength=0.5 * u.micron, efficiency=0.05, **kw):
+    """
+    Create a JWST telescope unit.
+
+    Parameters
+    ----------
+    wavelength : astropy.unit.Quantity
+        The wavelength at which it should be calculated.
+
+    R : float
+        The spectral resolution at which the
+        telescope will bin wavelengths..
+        (Ignored if telescope is None.)
+
+    dt : astropy.units.quantity.Quantity
+        The time over which the telescope exposes.
+        (Ignored if telescope is None.)
+    """
+    return define_telescope_unit(
+        telescope_name="Gemini",
+        wavelength=wavelength,
+        area=efficiency * 50 * u.m**2,
+        **kw,
+    )
+
 
 def define_HST_unit(wavelength=1.4 * u.micron, efficiency=0.05, **kw):
     """
@@ -220,6 +245,7 @@ telescope_units = dict(
     JWST=define_JWST_unit,
     APO=define_APO_unit,
     SBO=define_SBO_unit,
+    GMOS=define_GMOS_unit
 )
 
 
