### 2025-12-19 #9655 @editorialbot
[REVIEW]: exoatlas: friendly Python code for exoplanet populations
**Submitting author:** @zkbt (<a href="http://orcid.org/0000-0002-3321-4924">Zachory Berta-Thompson</a>)
**Repository:** https://github.com/zkbt/exoatlas
**Branch with paper.md** (empty if default branch):
**Version:** 0.7.0
