### 2026-01-02 #9655 @augustfly
@zkbt I have this as a joint submission with the AAS Journals (eJP ID AAS66890). Would you confirm that? And if it is true then can we label this is as "AAS" for tracking purposes @ivastar?  Thanks all.
