### 2026-01-03 #9655 @augustfly
It does @ivastar -- we have the manuscript as out for revision with @zkbt. Great!
