### 2026-01-03 #9655 @ivastar
I have added the AAS label.

@LorenzoMugnai, @tedjohnson12 please note that this paper is reviewed as part of our ongoing collaboration with [AAS publishing](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing) where authors submitting to one of the AAS journals can also publish a companion software paper in JOSS, thereby receiving a review of their software. For this service, JOSS receives a small donation from AAS publishing. Just making you aware. Please let me know if you have any questions or concerns about this.
