### 2026-01-19 #9655 @LorenzoMugnai
Thanks @ivastar, understood, and it makes sense that this paper should be assessed against the previous checklist.

That said, from what I can see so far, the repository already seems compatible with the current checklist as well (timeline, open development, collaborations, etc.), so I personally don’t see any harm in using the new list. The only differences I haven’t fully evaluated yet are on the paper side (software design aspects + AI-related points).

If you’re able to update the checklist on your side, that would be a huge help for me to track the review progress consistently. In the meantime, I’m also looking back at the checklist from a review I did in 2024, just to have a solid reference/trace.
