### 2026-01-19 #9655 @LorenzoMugnai
Hi @zkbt, congratulations on the work and effort that went into this package. I’m enjoying reading through it.

I am exploring the documentation: it looks clean and easy to follow. I still need to install the software and go through the instructions step by step, but overall, it is straightforward to navigate.

That said, I have two initial comments after a first pass. First, the documentation could benefit from a more complete landing page that more clearly states the problem the software aims to solve and how it addresses it, possibly including a simple figure/example plot to illustrate a typical use case and expected outputs. Second, I found the “Reference” section/menu slightly confusing at first glance, as it is not immediately clear what it contains. You may consider adding a short introduction/overview to that section to guide the reader.

I will keep posting my comments as I move forward with the review
