### 2026-01-19 #9655 @arfon
> @LorenzoMugnai due to changes in the JOSS scope what you are seeing is the new checklist. @warrickball can you please update the checklist or should I do it?

I can do it! (I did it).
