### 2026-01-20 zkbt/exoatlas#92 @tedjohnson12
Dependency issues related to `chromatic-lightcurve` and `setuptools`
`exoatlas` depends on [chromatic](https://github.com/zkbt/chromatic), which depends on the now deprecated `pkg_resources` (see [https://setuptools.pypa.io/en/latest/pkg_resources.html](https://setuptools.pypa.io/en/latest/pkg_resources.html)). This creates an error in fresh environments which don't include this package by default.

Three options to fix this are:
1. Update `chromatic` to list `setuptools` as a required dependency (see https://github.com/zkbt/chromatic/issues/271). Update `exoatlas` dependencies to require this new latest version of `chromatic`.
2. Update `chromatic` to use `importlib.resources`, as suggested by pypa (https://setuptools.pypa.io/en/latest/pkg_resources.html). Update `exoatlas` dependencies to require this new latest version of `chromatic`.
3. Update `exoatlas` list setuptools as a required dependency.

Option 3 is the quickest, option 2 is the "correct" way to do it, but any of these fixes will keep new users from giving up on using this really cool tool.

-----
More context:
In a fresh Python 14 environment, I did
```
$ pip install exoatlas
$ python
>>> import exoatlas as ea
... import exoatlas.visualizations as vi
... import astropy.units as u
...
... ea.version()
'0.7.1'
>>> nearby = exoplanets[exoplanets.distance() < 30 * u.pc]
... nearby.label = "Nearby"
... nearby.color = "coral"
... nearby.marker = "P"
... nearby.size = 20
>>> easy = exoplanets[
...     exoplanets.transmission_snr(telescope="JWST", wavelength=4 * u.micron) > 10
... ]
Traceback (most recent call last):
  File "<python-input-5>", line 2, in <module>
    exoplanets.transmission_snr(telescope="JWST", wavelength=4 * u.micron) > 10
...
File "/home/ted/anaconda3/envs/joss-exoatlas/lib/python3.14/site-packages/chromatic/imports.py", line 18, in <module>
    import copy, pkg_resources, os, glob, pickle
ModuleNotFoundError: No module named 'pkg_resources'
```
