### 2026-01-20 #9655 @tedjohnson12
Hey @zkbt! This is really an incredible tool you and your team have put together. I found it very easy to use and the visualizations it produces are beautiful. I ran into one minor challenge reproducing the quickstart portion of the docs, which I outlined in https://github.com/zkbt/exoatlas/issues/92 . I also share @LorenzoMugnai 's sentiment that the docs' landing page would benefit from being more filled-out. Otherwise, excellent work and I look forward to using this code!
