### 2026-01-26 #9655 @LorenzoMugnai
The manuscript is clear and well structured. However, I feel that the current discussion of the state of the field could be strengthened. In particular, the paper focuses on ExoAtlas, but does not sufficiently acknowledge other existing tools for planetary and exoplanet data visualisation and exploration. For example, resources such as [DACE](https://ui.adsabs.harvard.edu/abs/2019ASPC..521..757B/abstract) and similar platforms/tools represent alternative approaches within the community.

I believe the manuscript would benefit from a short paragraph that situates ExoAtlas within this broader ecosystem of tools. Specifically, it would be useful to briefly mention other available visualisation/data-exploration tools and clarify how ExoAtlas differs in scope, philosophy, or functionality (e.g., target audience, data integration, interactivity, mission focus, etc.). This would not require an extensive review, but would help readers better understand the unique contribution of the work in the context of existing community resources.
