### 2026-02-23 398a325
add `orbital_phase` keyword to `show_upcoming_transits` for secondary eclipses
