### 2026-02-23 742ca2f
hide tess_stars2px import inside function
