### 2026-03-14 c293894
fix bug to propagate keywords into plotted quantities in shoreline plots
