### 2026-03-15 7229442
tiny style changes + plotting improvements
