+++ exoatlas/visualizations/shorelines/preset_shorelines.py
@@ -223,7 +223,7 @@ class Shoreline_SemimajorAxis_x_StellarLuminosity_x_Radius(ShorelineErrorMap):
         log_P_2d = np.mean(samples_of_log_P_2d, axis=0)
 
         if colormesh:
-            background = plt.pcolormesh(
+            self._shoreline_probability_imshow = plt.pcolormesh(
                 10**log_a_2d,
                 10**log_L_2d,
                 log_P_2d,
@@ -234,7 +234,7 @@ class Shoreline_SemimajorAxis_x_StellarLuminosity_x_Radius(ShorelineErrorMap):
             )
 
         if contour:
-            plt.contour(
+            self._shoreline_probability_contours = plt.contour(
                 10**log_a_2d,
                 10**log_L_2d,
                 log_P_2d,
@@ -261,6 +261,9 @@ class Shoreline_SemimajorAxis_x_StellarLuminosity_x_Radius(ShorelineErrorMap):
         self.plot_hz_earth()
         self.plot_hz_kopparapu()
         self.plot_shoreline_probability()
+        plt.text(
+            0.5, 8, "habitable\nzone", va="top", ha="left", fontsize=8, color="seagreen"
+        )
 
 
 class ShorelineStandardMap(ShorelineErrorMap):
@@ -278,8 +281,8 @@ class ShorelineStandardMap(ShorelineErrorMap):
         L=LogRelativeStellarLuminosity,
     )
 
-    def __init__(self, order="vfL", **kw):
-        #
+    def __init__(self, order="vfL", posterior=None, **kw):
+        # decide which quantity is plotted on which axis
         x_var, y_var, slice_var = order
 
         # make sure they're all different
@@ -290,6 +293,7 @@ class ShorelineStandardMap(ShorelineErrorMap):
             xaxis=self.linear_plottables[x_var](lim=10 ** self.lims[x_var], **kw),
             yaxis=self.linear_plottables[y_var](lim=10 ** self.lims[y_var], **kw),
             sliceaxis=self.log_plottables[slice_var](lim=self.lims[slice_var], **kw),
+            posterior=posterior,
             **kw,
         )
 
@@ -352,7 +356,7 @@ class ShorelineStandardMap(ShorelineErrorMap):
 
         log_P_2d = np.mean(samples_of_log_P_2d, axis=0)
 
-        background = plt.pcolormesh(
+        self._shoreline_probability_imshow = plt.pcolormesh(
             10**log_x_2d,
             10**log_y_2d,
             log_P_2d,
@@ -361,7 +365,7 @@ class ShorelineStandardMap(ShorelineErrorMap):
             zorder=-1e9,
             rasterized=True,
         )
-        plt.contour(
+        self._shoreline_probability_contours = plt.contour(
             10**log_x_2d,
             10**log_y_2d,
             log_P_2d,
@@ -371,9 +375,73 @@ class ShorelineStandardMap(ShorelineErrorMap):
             colors=["gray", "black", "gray"],
         )
 
-    def refine(self, probability=True):
+    def label_flux_limits(
+        self, limits={"magma ocean": 1700 * u.K, "$\sf CO_2$ freezes": 194 * u.K}
+    ):
+
+        # figure out which axis is instellation
+
+        for k in self.plottable:
+            if isinstance(self.plottable[k], RelativeInstellation):
+                instellation_axis = k
+
+        try:
+            instellation_axis
+        except NameError:
+            raise NameError(
+                "This map doesn't seem to have any instellation axis,"
+                "so we can't plot flux limits on it."
+            )
+
+        # calculate flux limits
+        def calculate_S(T, f=1 / 4, A=0):
[…]
