### 2026-03-18 04cc05e
additions for revision of cosmic shoreline paper
