+++ exoatlas/visualizations/shorelines/preset_shorelines.py
@@ -53,7 +53,9 @@ class Shoreline_SemimajorAxis_x_StellarLuminosity_x_Radius(ShorelineErrorMap):
         relative_escape_velocity = self._convert_radius_to_relative_escape_velocity(R)
         return relative_escape_velocity
 
-    def plot_hz_earth(self, color="seagreen", alpha=0.75, ax=None, **kw):
+    def plot_hz_earth(
+        self, color="seagreen", alpha=0.75, ax=None, annotate=False, **kw
+    ):
         """
         Plot the Earth equivalent habitable zone.
 
@@ -71,6 +73,17 @@ class Shoreline_SemimajorAxis_x_StellarLuminosity_x_Radius(ShorelineErrorMap):
         # plot Earth-equivalent instellation distance
         a_ee = 1 * u.AU * (L / u.Lsun) ** 0.5
         plt.plot(a_ee, L, color=color, alpha=alpha, **kw)
+
+        if annotate:
+            plt.text(
+                1.8,
+                7,
+                "habitable\nzone   ",
+                va="top",
+                ha="right",
+                fontsize=7,
+                color="seagreen",
+            )
 
     def plot_hz_kopparapu(
         self,
@@ -132,12 +145,12 @@ class Shoreline_SemimajorAxis_x_StellarLuminosity_x_Radius(ShorelineErrorMap):
 
         if annotate:
             plt.text(
-                np.max(a_inner),
-                np.max(L) * 0.8,
-                "habitable zone",
+                1.8,
+                7,
+                "habitable\nzone   ",
                 va="top",
                 ha="right",
-                fontsize=8,
+                fontsize=7,
                 color="seagreen",
             )
 
@@ -248,28 +261,22 @@ class Shoreline_SemimajorAxis_x_StellarLuminosity_x_Radius(ShorelineErrorMap):
             a_text = 10 ** np.interp(0.5, log_P_2d[-1, :], log_a_1d)
             plt.text(
                 a_text,
-                10 ** np.max(log_L_1d) * 0.8,
+                7,
                 "   cosmic\n    shoreline",
                 va="top",
                 ha="left",
-                fontsize=8,
+                fontsize=7,
                 color="black",
                 alpha=0.5,
             )
 
-    def refine(self):
-        self.plot_hz_earth()
+    def refine(self, **kw):
+        self.plot_hz_earth(annotate=True)
         self.plot_hz_kopparapu()
-        self.plot_shoreline_probability()
-        plt.text(
-            0.5, 8, "habitable\nzone", va="top", ha="left", fontsize=8, color="seagreen"
-        )
+        self.plot_shoreline_probability(annotate=True)
 
 
 class ShorelineStandardMap(ShorelineErrorMap):
-    lims = dict(
-        v=np.array([-1.5, 1.0]), f=np.array([-4.0, 4.0]), L=np.array([0.5, -3.5])
-    )
 
     linear_plottables = dict(
         v=RelativeEscapeVelocity, f=RelativeInstellation, L=RelativeStellarLuminosity
@@ -281,12 +288,26 @@ class ShorelineStandardMap(ShorelineErrorMap):
         L=LogRelativeStellarLuminosity,
     )
 
-    def __init__(self, order="vfL", posterior=None, **kw):
+    def __init__(
+        self, order="vfL", posterior=None, vlim=None, flim=None, Llim=None, **kw
+    ):
         # decide which quantity is plotted on which axis
         x_var, y_var, slice_var = order
 
         # make sure they're all different
         assert sorted(order) == ["L", "f", "v"]
+
+        # set default limits
+        self.lims = dict(
+            v=np.array([-1.5, 1.0]), f=np.array([-4.0, 4.0]), L=np.array([0.5, -3.5])
[…]
