+++ exoatlas/version.py
@@ -1,4 +1,4 @@
-__version__ = "0.7.4"
+__version__ = "0.7.5"
 
 
 def version():
