### 2026-03-24 #9655 @augustfly
Just a ping that the corresponding AAS manuscript was just accepted today.
