### 2026-05-07 4d6f276
auto-load posterior from zenodo
