### 2026-05-11 d0348fd
allow kludged radii for better imaging constrast plots
