+++ exoatlas/version.py
@@ -1,4 +1,4 @@
-__version__ = "0.7.5"
+__version__ = "0.7.6"
 
 
 def version():
