### 2026-05-15 7ffea2a
refactor, moving `calculations/` up for easier shoreline imports
