+++ exoatlas/calculations/stellar.py
@@ -1,4 +1,4 @@
-from ...imports import *
+from ..imports import *
 
 
 '''def stellar_luminosity_from_table(self, distribution=False, **kw):
