### 2026-05-18 8e6a363
add shoreline documentation
