+++ exoatlas/calculations/shoreline.py
@@ -11,7 +11,7 @@ class Shoreline:
 
     var_names = ["log_f_0", "p", "q", "ln_w"]
 
-    def __init__(self, posterior=None, **kw):
+    def __init__(self, posterior=None, cache=True, **kw):
         """
         Initialize a Shoreline object.
 
@@ -24,6 +24,14 @@ class Shoreline:
             If None, the default posterior will be downloaded
             from the Zenodo repository associated with the
             Berta-Thompson et al. (2026) Cosmic Shoreline.
+        cache : bool
+            If no posterior is provided and the posterior 
+            is being downloaded from Zenodo, this sets whether
+            `download_file` should cache a copy of the file 
+            locally to avoid re-downloading it every time 
+            the shoreline is needed. This should mostly be 
+            True, unless for some reason the local copy of 
+            the posteriors needs to be updated.
         **kw : dict
             All additional keywords will be ignored.
         """
@@ -36,10 +44,7 @@ class Shoreline:
             filename = posterior
             self.posterior = az.from_netcdf(filename)
         elif posterior is None:
-            # download the default posterior from Zenodo
-            link = "https://zenodo.org/records/15858798/files/cosmic-shoreline-btwm2026-posterior.nc?download=1"
-            filename = download_file(link)
-            self.posterior = az.from_netcdf(filename)
+            self.posterior = self.download_posterior(cache=cache)
         else:
             # nothing else
             raise NotImplementedError(
@@ -54,6 +59,28 @@ class Shoreline:
 
     def __repr__(self):
         return f"<🏝️{self.var_names}>"
+
+    def download_posterior(self, cache=True):
+        ''' 
+        Download the latest version of the BTMW26 posterior from Zenodo.
+
+        Parameters 
+        ----------
+        cache : bool
+            Should `download_file` cache a copy of the file 
+            locally to avoid re-downloading it every time 
+            the shoreline is needed? This should mostly be 
+            True, unless for some reason the local copy of 
+            the posteriors needs to be updated.
+        '''
+        # define the link
+        zenodo_code = '15858798'
+        posterior_filename = 'cosmic-shoreline-btwm2026-posterior.nc'
+        link = f"https://zenodo.org/records/{zenodo_code}/files/{posterior_filename}?download=1"
+
+        # use astropy to download the file
+        filename = download_file(link, cache=cache)
+        return az.from_netcdf(filename)
 
     def best_parameters(self):
         """
@@ -101,7 +128,7 @@ class Shoreline:
         return log_f_0 + p * log_v + q * log_L
 
     def probability_of_atmosphere(
-        self, log_f_0=1.0, p=4.0, q=0.0, ln_w=0, log_v=0, log_L=0, log_f=0
+        self, log_f_0=1.0, p=4.0, q=0.0, ln_w=0, log_v=0, log_L=0, log_f=0, **kw
     ):
         """
         Calculate the probability a planet has an atmosphere.
@@ -128,6 +155,8 @@ class Shoreline:
             (independent variable), log10(stellar luminosity relative to Sun)
         log_f : float, Quantity, array
             (independent variable), log10(bolometric flux relative to Earth)
+        **kw : dict 
+            All additional keywords will be ignored.
         """
         distance_from_shoreline = log_f - self.log_f_shoreline(
             log_f_0=log_f_0, p=p, q=q, log_v=log_v, log_L=log_L
@@ -220,13 +249,13 @@ class Shoreline:
         if latex:
             return [
                 f"{latexify_confidence_interval(m*100, l*100, u*100)}\%"
-                for m, l, u in zip(median, lower, upper)
+                for m, l, u in zip(np.atleast_1d(median), np.atleast_1d(lower), np.atleast_1d(upper))
[…]
