### 2026-05-18 5ccaa27
add better shoreline test
