+++ exoatlas/tests/test_shoreline.py
@@ -1,17 +1,13 @@
 from .setup_tests import *
 from exoatlas import *
+from exoatlas.visualizations import * 
 from exoatlas.calculations.shoreline import *
 
 
-def test_probability_of_atmosphere():
-    # need to change for other people to be able to run tests
-    posterior_filename = "/Users/zabe0091/Dropbox/zach/code/exoatlas/sandbox/shoreline/all-any-uncertainties=True-numpyro.nc"
+def test_shoreline_probability_of_atmosphere():
     # need to add automatic download for posteriors!
     s = SolarSystem()
-    import arviz as az
-
-    posterior = az.from_netcdf(posterior_filename)
-    shoreline = Shoreline(posterior)
+    shoreline = Shoreline()
 
     p = s.probability_of_atmosphere(shoreline=shoreline)
     p_uncertainty = s.probability_of_atmosphere_uncertainty(shoreline=shoreline)
@@ -23,3 +19,31 @@ def test_probability_of_atmosphere():
     plt.xscale("log")
     plt.xlabel("Planet Semimajor Axis (AU)")
     plt.ylabel("Probability of Atmosphere")
+
+
+def test_shoreline_visualizations():
+
+    # create populations
+    s = SolarSystem()
+    e = TransitingExoplanets()
+
+    # pick the name of the planet to highlight
+    planet_name = 'LTT1445Ab'
+
+    # create a subset population to highlight that one planet
+    highlight = e[planet_name]
+
+    # define a single panel of the visualization
+    m = ShorelineStandardMap()
+
+    # construct a grid of multiple slices
+    g = SliceGridGallery(m, N=4)
+
+    # add the planets to the panels
+    g.build([s, e, highlight])
+
+    # add embellishments to the plot, including probability 
+    g.refine()
+
+    # colorbar for probability
+    g.add_colorbar()
