### 2026-05-18 d7fd18f
add a little more ecosystem context
