+++ paper/paper.bib
@@ -1,19 +1,82 @@
+@misc{eyes,
+  author = {NASA},
+  title = {NASA's Eyes},
+  year = {2026},
+  url = {https://science.nasa.gov/eyes/},
+  note = {Accessed: 2026-05-18}
+}
+
+@article{stellarium,
+  title = {Simulated {{Sky}}: {{Stellarium}} for {{Cultural Astronomy Research}}},
+  shorttitle = {Simulated {{Sky}}},
+  author = {Zotti, Georg and Hoffmann, Susanne M. and Wolf, Alexander and Ch{\'e}reau, Fabien and Ch{\'e}reau, Guillaume},
+  year = 2021,
+  month = mar,
+  journal = {Journal of Skyscape Archaeology},
+  volume = {6},
+  number = {2},
+  issn = {2055-3498, 2055-348X},
+  doi = {10.1558/jsa.17822},
+  urldate = {2026-05-18},
+}
+
+
+@INPROCEEDINGS{dace,
+       author = {{Buchschacher}, Nicolas and {Alesina}, Fabien},
+        title = "{DACE: New Available Visualisation and Analysis Tools for Exoplanet Research}",
+    booktitle = {Astronomical Data Analysis Software and Systems XXVI},
+         year = 2019,
+       editor = {{Molinaro}, Marco and {Shortridge}, Keith and {Pasian}, Fabio},
+       series = {Astronomical Society of the Pacific Conference Series},
+       volume = {521},
+        month = oct,
+        pages = {757},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ASPC..521..757B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@software{tapir,
+       author = {{Jensen}, Eric},
+        title = "{Tapir: A web interface for transit/eclipse observability}",
+ howpublished = {Astrophysics Source Code Library, record ascl:1306.007},
+         year = 2013,
+        month = jun,
+          eid = {ascl:1306.007},
+archivePrefix = {ascl},
+       eprint = {1306.007},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2013ascl.soft06007J},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@misc{varastro,
+  author = {Czech Astronomical Society},
+  title = {VarAstro,
+  year = {2026},
+  url = {https://var.astro.cz/en},
+  note = {Accessed: 2026-05-18}
+}
+
+
 @ARTICLE{exo-archive,
        author = {{Christiansen}, Jessie L. and {McElroy}, Douglas L. and {Harbut}, Marcy and {Ciardi}, David R. and {Crane}, Megan and {Good}, John and {Hardegree-Ullman}, Kevin K. and {Kesseli}, Aurora Y. and {Lund}, Michael B. and {Lynn}, Meca and {Muthiar}, Ananda and {Nilsson}, Ricky and {Oluyide}, Toba and {Papin}, Michael and {Rivera}, Amalia and {Swain}, Melanie and {Susemiehl}, Nicholas D. and {Tam}, Raymond and {van Eyken}, Julian and {Beichman}, Charles},
         title = "{The NASA Exoplanet Archive and Exoplanet Follow-up Observing Program: Data, Tools, and Usage}",
-      journal = {arXiv e-prints},
-     keywords = {Earth and Planetary Astrophysics, Instrumentation and Methods for Astrophysics, Solar and Stellar Astrophysics},
+      journal = {\psj},
[…]
