### 2026-05-18 58c9890
fix typo in .bib
