+++ paper/paper.bib
@@ -50,7 +50,7 @@ archivePrefix = {ascl},
 
 @misc{varastro,
   author = {Czech Astronomical Society},
-  title = {VarAstro,
+  title = {VarAstro},
   year = {2026},
   url = {https://var.astro.cz/en},
   note = {Accessed: 2026-05-18}
