### 2026-05-19 a211f18
update dependencies to fix old `pkg_resources` problem in `chromatic`
