+++ setup.py
@@ -85,7 +85,7 @@ setup(
         "astropy>=3.2.3",
         "astroquery>=0.3.9",
         "astroplan>=0.10",
-        "rainbow-connection>=0.0.1",
+        "rainbow-connection>=0.0.17",
         "PyYAML",
         "tqdm",
         "pytz",
