### 2026-05-19 zkbt/exoplanet-atlas#92 @zkbt
Thanks, @tedjohnson12 ! Sorry for the delay on this, but here's what we've done for the fix.

- `chromatic-lightcurves` version 0.5.1 replaces the deprecated `pkg_resources` with `importlib.resources`
- `rainbow-connection` version 0.0.17 now includes `chromatic-lightcurves>=0.5.1` as a dependency
- `exoatlas` will very soon include `rainbow-connection>=0.0.17` as a dependency.

#103 implements this change and will be merged in and released soon
