### 2026-05-22 45110bf
improving documentation + API
