+++ exoatlas/populations/pineda_skew.py
@@ -378,8 +378,10 @@ def make_skew_samples_from_lowerupper(
 
     Returns
     -------
-
-
+    samples : Quantity and/or array-like
+        an array with shape (N_data, N_samples) with samples drawn from 
+        a skew normal distribution that matches the provided center and 
+        asymmetric lower/upper limits
     """
 
     assert np.max([len(np.shape(x)) for x in [mu, sigma_lower, sigma_upper]]) <= 1
