+++ exoatlas/calculations/planetary.py
@@ -147,9 +147,9 @@ def eccentricity(self, distribution=False, **kw):
 
 def argument_of_periastron(self, distribution=False, **kw):
     """
-    Planet Orbital Argument of Periastron ($\omega$, degrees)
+    Planet Orbital Argument of Periastron (omega, degrees)
 
-    The argument of periastron of the planet's orbit, $\omega$.
+    The argument of periastron of the planet's orbit (omega).
     If it is `nan` (often because eccentricity is zero), then
     this will be assumed to be 0.
 

[…]
