### 2026-05-22 6e309b6
tiny doc tweaks
