+++ setup.py
@@ -86,12 +86,12 @@ setup(
         "astroquery>=0.3.9",
         "astroplan>=0.10",
         "rainbow-connection>=0.0.17",
-        "PyYAML",
-        "tqdm",
-        "pytz",
-        "tess-point",
+        "tqdm>=4.67.1",
+        "pytz>2025.2",
+        "tess-point>0.9.2",
+        "arviz>=0.21.0",
         "adjusttext",
-        "arviz",
+        #"PyYAML",
     ],
     extras_require={
         "develop": [
@@ -100,13 +100,13 @@ setup(
             "black[jupyter]",
             "jupyter",
             "ipython",
-            "mkdocs",
-            "mkdocs-material",
-            "mkdocstrings",
-            "mkdocstrings-python",
+            "mkdocs>=1.6.1",
+            "mkdocs-material>=9.6.9",
+            "mkdocstrings>=0.29.0",
+            "mkdocstrings-python>=1.16.7",
             "pytkdocs[numpy-style]",
-            "mkdocs-jupyter>=0.26.0",
-            "mkdocs-exclude",
+            "mkdocs-jupyter>=0.26.3",
+            "mkdocs-exclude>=1.0.2",
             "twine",
             "pre-commit",
         ]
