+++ docs/developer/index.md
@@ -1,3 +1,3 @@
-# Developer Guide
+# 🤓 Developer Guide
 
 This developer guide is meant to help you figure out how to contribute something new to `exoatlas`. Please consider helping improve `exoatlas`, whether that means reporting a bug, suggesting a new feature, or writing oodles of neat code.
