+++ docs/api_index.md
@@ -1,3 +1,3 @@
-# API Reference 
+# 📚 API Reference 
 
 This section provides a detailed reference for the definitions of key classes and functions in `exoatlas`. It collects all the docstrings from all populations, visualizations, and helper functions into one location. We include this here mostly to be a searchable reference for the nitty gritty of how `exoatlas` works; it's probably the last place you should look to try to learn the basics of how `exoatlas` populations and visualizations work. Instead, please try the [User Guide](user/index.md).
