+++ docs/user/index.md
@@ -1,3 +1,3 @@
-# User Guide
+# 🤩 User Guide
 
-This user guide is meant to help you figure out how to use `exoatlas` for research, teaching, and learning. If something in here seems unclear, please [open an Issue](../developer/github.ipynb) so we can help improve it for everyone!
+This user guide is meant to help you figure out how to use `exoatlas` for research, teaching, and learning. Each section in the menu is self-contained, so feel free to skip around as needed.
