### 2026-05-30 ef418bc
fix shoreline to work with both old and new arviz
