+++ mkdocs.yml
@@ -71,13 +71,14 @@ plugins:
         - "*.npy"
         - "*.ecsv"
 markdown_extensions:
-    - toc:
-        permalink: "#"
-        toc_depth: 1
-    - pymdownx.arithmatex:
-        generic: true
+  - toc:
+      permalink: "#"
+      toc_depth: 1
+  - pymdownx.arithmatex:
+      generic: true
 extra_javascript:
   - javascripts/mathjax.js
+  - https://unpkg.com
 
 # this is super borrowed from Christina Hedges' fabulous
 # https://christinahedges.github.io/astronomy_workflow/
