### 2026-05-30 f660b3c
tiny animation tweaks
