### 2026-05-30 zkbt/exoplanet-atlas#92 @zkbt
I'm calling this closed with today's release of version 0.7.7.
