+++ paper/paper.md
@@ -66,9 +66,9 @@ Planets are complicated. Understanding how they work requires connecting individ
 - planning future telescope observations of known exoplanet systems
 - making beautiful and up-to-date planet data visualizations
 
-Online planetary data archives merge incredible curatorial efforts with powerful tools for data access and visualization. These archives include the NASA Exoplanet Archive [@exo-archive], exoplanet.eu [@exoplaneteu], exo.MAST [@exomast], the University of Geneva Data Analysis Center for Exoplanets [@dace], TEPCat [@tepcat], and the Open Exoplanet Catalog [@open-exo] for exoplanets, and the JPL Solar System Dynamics [@ssd], IAU Minor Planet Center [@mpc], the NSSDCA Planetary Facts sheets [@facts] for Solar System objects. Through web portals, many of these archives enable the user to find or quickly create up-to-date plots of planetary properties. Through downloadable tables and/or application programming interfaces (APIs), they can provide raw or curated planetary properties with clear citations to the scientific literature. Drawing from these and other archives, tools like NASA's Eyes [@eyes] and Stellarium [@stellarium] allow the easy visualization of planetary orbits, and tools like Tapir [@tapir], the Exoplanet Archive Transit and Ephemeris Service [@exo-archive], and VarAstro [@varastro] help plan new exoplanet observations.
+Online planetary data archives merge incredible curatorial efforts with powerful tools for data access and visualization. These archives include the NASA Exoplanet Archive [@exo-archive], exoplanet.eu [@exoplaneteu], exo.MAST [@exomast], the University of Geneva Data Analysis Center for Exoplanets [@dace], TEPCat [@tepcat], and the Open Exoplanet Catalog [@open-exo] for exoplanets, and the JPL Solar System Dynamics [@ssd], IAU Minor Planet Center [@mpc], and the NSSDCA Planetary Facts sheets [@facts] for Solar System objects. Through web portals, many of these archives enable the user to find or quickly create up-to-date plots of planetary properties. Through downloadable tables and/or application programming interfaces (APIs), they can provide raw or curated planetary properties with clear citations to the scientific literature. Drawing from these and other archives, tools like NASA's Eyes [@eyes] and Stellarium [@stellarium] allow the easy visualization of planetary orbits, and tools like Tapir [@tapir], the Exoplanet Archive Transit and Ephemeris Service [@exo-archive], and VarAstro [@varastro] help plan new exoplanet observations.
 
-`exoatlas` does not intend to replace any of these important archival efforts or their web plotting portals. The main gap that `exoatlas` aims to address is to support non-standard calculations and creative visualizations by users who are fluent in Python but uninterested in carefully parsing complicated tables, calculating replacements for missing quantities, or manually propagating uncertainties into derived quantities. By trying to do a good job of these steps behind the scenes, `exoatlas` aims to provide an approachable interface for exploratory analysis, centering a strong focus on scientific storytelling and illuminating visualizations. The target audience is the scientist, educator, or student who wants to curiously explore one step beyond a web plot, to playfully interact with exoplanet and Solar System data in Python, to fussily adjust how data are filtered or plotted, and to easily integrate new planet discoveries into their workflow.
[…]
