### 2026-09-03 #9655 @zkbt
@ivastar , can I please check in on the timeline for this? Should ApJ proceed without us?
