### 2026-09-04 #9655 @warrickball
Hi @zkbt, I apologies that things have been busy and many editors themselves unavailable over the summer. While we wait for @ivastar to step back in, I'll let you know that the DOI of the JOSS article is determined by the review number and will be `10.21105/joss.09655`.

If it's volume and issue number you need, I'll need to confirm but IIRC the volume is set by the year and currently 11. I think the issue number is by month and will be 125 if we complete the publication process this month.

If that's enough information, AAS can proceed without us. Once you know the AAS DOI, you can also add it to the relevant metadata field in your JOSS `paper.md`. (_Edit_: I only just noticed that the DOI is known above, so if you haven't already, you can add it to the JOSS paper.)
