### 2026-09-04 #9655 @warrickball
Just looking ahead in our process, @LorenzoMugnai, can I confirm it's OK to tick off everything in [your review checklist](https://github.com/openjournals/joss-reviews/issues/9655#issuecomment-3767355728)?
