### 2026-09-04 #9655 @zkbt
@warrickball , no need to apologize at all! I passed this on to the AAS, and I just added the ApJ DOI to the `paper.md`. Thanks! Please let me know if there's anything else I can do to help!
